requirements.txt
arabic-reshaper==3.0.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
comm==0.1.2
contextlib2==21.6.0
contourpy==1.1.0
creativeai==0.1.1
cryptography==39.0.1
ctcdecode==1.0.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
demoji==1.1.0
docopt==0.6.2
executing==1.2.0
fastrlock==0.8.1
fonttools==4.39.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.0.0
importlib-resources==6.0.0
ipykernel==6.21.2
ipython==8.11.0
ipywidgets==8.0.4
jedi==0.18.2
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langid==1.1.6
lazy-loader==0.3
lxml==4.9.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
progressbar2==3.55.0
prompt-toolkit==3.0.38
protobuf==4.22.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.4.22
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pytest==2.6.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-utils==3.5.2
PyWavelets==1.4.1
pyzmq==25.0.0
requests==2.28.2
schema==0.7.5
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
texturize==0.12
tifffile==2023.7.10
torch==1.13.1
torchvision==0.14.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.5.0
uno==0.3.3
unotools==0.3.3
urllib3==1.26.14
wcwidth==0.2.6
widgetsnbextension==4.0.5
zipp==3.15.0
